Language Services/language_analysis_app/requirements.txt
flask
azure-ai-textanalytics
python-dotenv
